Stable Lobot Refactor
Also replaces masthead with new webp to check quality drop in image on production.
