+++ lobot/css/javascript.css
@@ -1,0 +1,1 @@
+.about-section {}
