+++ lobot/css/teamGrid.css
@@ -1,0 +1,99 @@
+.team-grid {
+    color:#313437;
+    background-color:#fff;
+}
+
+.team-grid p {
+    color:#7d8285;
+}
+
+.team-grid h2 {
+    font-weight:bold;
+    margin-bottom:40px;
+    padding-top:40px;
+    color:inherit;
+}
+
+@media (max-width:767px) {
+    .team-grid h2 {
+        margin-bottom:25px;
+        padding-top:25px;
+        font-size:24px;
+    }
+}
+
+.team-grid .intro {
+    font-size:16px;
+    max-width:500px;
+    margin:0 auto;
+}
+
+.team-grid .intro p {
+    margin-bottom:0;
+}
+
+.team-grid .people {
+    padding:50px 0;
+}
+
+.team-grid .item {
+    margin-bottom:30px;
+}
+
+.team-grid .item .box {
+    text-align:center;
+    background-repeat:no-repeat;
+    background-size:cover;
+    background-position:center;
+    height:300px;
+    position:relative;
+    overflow:hidden;
+}
+
+.team-grid .item .cover {
+    position:absolute;
+    top:0;
+    left:0;
+    width:100%;
+    height:100%;
+    background-color: #FAA300;
+    transition:opacity 0.15s ease-in;
+    opacity:0;
+    padding-top:80px;
+    color:#000000;
+    text-shadow:1px 1px 1px rgba(0,0,0,0.15);
+}
+
+.team-grid .item:hover .cover {
+    opacity:1;
+}
+
+.team-grid .item .name {
+    font-weight:bold;
+    margin-bottom:8px;
+}
+
+.team-grid .item .title {
+    text-transform:uppercase;
+    font-weight:bold;
+    color:#495057;
+    letter-spacing:2px;
+    font-size:13px;
+    margin-bottom:20px;
+}
+
+.team-grid .social {
+    font-size:18px;
+}
+
+.team-grid .social a {
+    color:inherit;
+    margin:0 10px;
+    display:inline-block;
+    opacity:0.7;
+}
+
+.team-grid .social a:hover {
+    opacity:1;
+}
+
