+++ lobot/lobot.js
@@ -1,29 +1,36 @@
-import { exists} from "https://deno.land/std/fs/mod.ts";
-import { HtmlHead } from "../components/head.js";
+import { HtmlHead } from "./html/head.js";
 
 const componentBundle = {
     HtmlHead: HtmlHead,
 }
 
+/** Reads and returns the text content of the given resource **/
 async function LoadResource (name, type) {
-  const lobotResourcePath = `lobot/resources/${name}.${type}`;
-  const islobotResource = await exists(lobotResourcePath);
-  if (islobotResource) {
-    return "css" ? `<style>${await Deno.readTextFile(lobotResourcePath)}</style>` : `<script>${await Deno.readTextFile(lobotResourcePath)}</script>`;
-  }
+  const lobotResourcePath = `lobot/${type}/${name}.${type}`;
+
+  return (type === `css`)
+          ? `<style>${await Deno.readTextFile(lobotResourcePath)}</style>`
+          : (type === `js`)
+          ? `<script>${await Deno.readTextFile(lobotResourcePath)}</script>`
+          : (type === `txt`)
+          ? `<p>${await Deno.readTextFile(lobotResourcePath)}</p>`
+          : (type === `json`)
+          ? await Deno.readTextFile(lobotResourcePath)
+          : null;
 }
 
-async function LoadComponent () {
+async function LoadHtmlBundle () {
   return componentBundle
 }
 
 async function Serve(resource) {
-  const name = resource.split(":")[0];
-  const type = resource.split(":")[1];
-  if (type === "css" || type === "js") {
+  const name = resource.split(`:`)[0];
+  const type = resource.split(`:`)[1];
+  if (type === `css` || type === `js` || type === `txt` ) {
     return await LoadResource(name, type);
-  } else {
-    return await LoadComponent(name);
+  }
+  if (type === `html`) {
+    return await LoadHtmlBundle(name);
   }
 }
